# docs/swagger.yaml
definitions:
  dto.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  dto.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/dto.ErrorDetail'
    type: object
...
  dto.PullRequestShortDTO:
    properties:
      author_id:
        type: string
      pull_request_id:
        type: string
      pull_request_name:
        type: string
      status:
        type: string
    type: object
...
  dto.UserDTO:
    properties:
      is_active:
        type: boolean
      team_name:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
  request.BatchDeactivateTeamRequest:
    properties:
      team_name:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - team_name
    type: object
  request.BatchDeactivateUsersRequest:
    properties:
      user_ids:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - user_ids
    type: object
...
  request.SetUserActiveRequest:
    properties:
      is_active:
        type: boolean
      user_id:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - user_id
    type: object
...
  response.AllUsersResponse:
    properties:
      count:
        type: integer
      users:
        items:
          $ref: '#/definitions/dto.UserDTO'
        type: array
    type: object
  response.BatchDeactivateResponse:
    properties:
      deactivated_users:
        items:
          type: string
        type: array
      processing_time_ms:
        type: integer
      reassigned_prs:
        items:
          $ref: '#/definitions/response.PRReassignmentInfo'
        type: array
      skipped_users:
        items:
          type: string
        type: array
      total_deactivated:
        type: integer
      total_prs_reassigned:
        type: integer
    type: object
...
  response.PRReassignmentInfo:
    properties:
      new_reviewers:
        items:
          type: string
        type: array
      old_reviewers:
        items:
          type: string
        type: array
      pull_request_id:
        type: string
    type: object
...
  response.UserResponse:
    properties:
      user:
        $ref: '#/definitions/dto.UserDTO'
    type: object
  response.UserReviewsResponse:
    properties:
      pull_requests:
        items:
          $ref: '#/definitions/dto.PullRequestShortDTO'
        type: array
      user_id:
        type: string
    type: object
info:
  contact: {}
  description: Service for automatic PR reviewer assignment
  title: PR Reviewer Assignment Service API
  version: "1.0"
paths:
...
  /admin/users:
    get:
      consumes:
      - application/json
      description: Get list of all users in the system
      produces:
      - application/json
      responses:
        "200":
          description: Users retrieved successfully
          schema:
            $ref: '#/definitions/response.AllUsersResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden - Admin access required
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all users (Admin only)
      tags:
      - Users
...
  /users/batchDeactivateTeam:
    post:
      consumes:
      - application/json
      description: Deactivate all members of a team and safely reassign their open
        PRs
      parameters:
      - description: Batch deactivate team request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.BatchDeactivateTeamRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Team members deactivated and PRs reassigned
          schema:
            $ref: '#/definitions/response.BatchDeactivateResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden - Admin access required or cannot deactivate admin
            team
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Batch deactivate team members (Admin only)
      tags:
      - Users
  /users/batchDeactivateUsers:
    post:
      consumes:
      - application/json
      description: Deactivate specified users and safely reassign their open PRs
      parameters:
      - description: Batch deactivate users request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.BatchDeactivateUsersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Users deactivated and PRs reassigned
          schema:
            $ref: '#/definitions/response.BatchDeactivateResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden - Admin access required
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Batch deactivate users (Admin only)
      tags:
      - Users
  /users/getReview:
    get:
      consumes:
      - application/json
      description: Get list of pull requests where user is assigned as reviewer
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User reviews retrieved successfully
          schema:
            $ref: '#/definitions/response.UserReviewsResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get PRs assigned to user
      tags:
      - Users
  /users/setIsActive:
    post:
      consumes:
      - application/json
      description: Update user's active status. Admin users cannot be deactivated.
      parameters:
      - description: Set active request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.SetUserActiveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User status updated successfully
          schema:
            $ref: '#/definitions/response.UserResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden - Admin access required or cannot deactivate admin
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Set user active status (Admin only)
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"